Avoid infinite loop on account verification. #43
When activation link of an already active account is visited, the
message shown has a link to login. By default, the login page redirects
to the referring page after successful login. So after login, it is
coming back to actication page with link to login, looking like an
infinite loop.

Added `?redirect=/` to login url to tell login page to redirect to home
page after successful login.