import_authors fixes to support multiple authors
Apply a patch from Noufal that fixes the import API to better handle
multiple authors. The import code was written for loading MARC, the
MARC standard only allows for a single author. In MARC if there are
multiple co-authors then the first author is listed as the author and
the others are listed as contributors.