Redirect ibsn requests to www.archive.org whenever possible.
The covers are stored in archive.org as zip files with 10K items in each
zip file. Each item contains 4 zip files: containing original, small, medium and
large covers respectively.

When a cover is requested by isbn, after finding the coverid,
archive.org url is constructed to access the image in the zip file and
redirected to that url.