Implemented /books/ia:foo00bar urls.
When the URL is accesses, the archive.org metadata API is queried to get
the metadata of foo00bar item and it is displayed as book record on
openlibrary. Borowing also works for these records.

Now we don't have to worry about import not working well etc. Once we
add these records to search engine, they show up in the search results
and work like first class records.

Haven't figured out how to handle edits on these pages.

Todos:
* show 404 when the item is not found on archive.org or it is not a text item
* redirect if an OL record alredy exists with the identifier as ocaid