Separate and de-duplicate CSS for #editions component
For readability we move this code to a single file.

In terms of impact on the site this should be a NOOP - the same CSS
will be shipped as before

By using nesting, we can also get a clearer idea of the hierarchy of
these components, which helps find redundant rules

Many of the rules in datatables duplicate the rules in master so have
been removed.