Don't allow duplicate selectors and properties
Quite a problem common in managing CSS is several selectors which match
the exact same thing and worse still duplicate CSS properties - the
end result is the first property gets ignored and its not clear when
certain CSS selectors later on in the file override your previous
rule.

We can detect these and fix them.

This surfaces several rules that are being sent to the user but are
not needed and certain cases where the wrong css property is being
used - .sorting_desc_disabled was using background-position instead
of background-image for instance.

W00t!