Possible to load styles via JavaScript
Add a new JavaScript function loadScript - this will allow us
to add styles to the page via JavaScript.

The importance of such a function can be seen with this simple example
- the books and editions editor both make use of jquery autocomplete
but no other pages use it.

We can thus limit the loading of this CSS to the books 'edit' page.
Other pages will benefit from not loading the unnecessary CSS.
We will use this technique later for various other pages.

Note: the jquery autocomplete JavaScript will continue to load on
all pages. This is not an oversight, but right I focus solely
on the CSS.

Fixes: #1094