Slick and Font awesome code loaded in separate stylesheet
Long term we need to move these stylesheets off the critical path. They are only
needed when JavaScript is enabled so do not need to be render
blocking. If JS is disabled, users get penalised with bad performance.

It's not that easy as doing so now would add some highly visible flashes of
unstyled content and we haven't started tackling the JS
so we must be patient. For now, however we can
reduce the 3 http requests for these files into 1.

Other changes:
* A new lib folder is added to the static/css folder to track external
libraries that have been pulled in.
* Paths to fonts and ajax loader are now absolute URIs