Remove some link styles back to legacy
On closer inspection these are not needed on the home page and
possibly unnecessary. The "on" class only seems to appear when
used with links inside a ".tools" element and then only when
JavaScript is enabled.

Since it's only added with JavaScript (
inside openlibrary/plugins/openlibrary/js/ol.js) move to js-all
entry point.