Load carousel-JavaScript specific styles via JavaScript
Many of the carousel styles are not needed until the carousel has
been fully initialised.

Since slick-theme is loaded as part of the js-all entry point, some
of these styles become mismatched causing a jarring repaint/reflow
of the display.

This moves those styles out of the critical path, which lowers the
initial CSS footprint and also progressively enhances the non-js
carousel to the JavaScript one with a neligible flash

A small re-adjustment is required given that the carousel sets width
on the carousel__items

Fixes: #1173