Author autocomplete added to "books/add" view
DRY up the author autocomplete functionality so it can be added
anywhere and add it to the add route.

Remove the js-books-edit entry point and move the CSS for the
autocomplete into the global js-all file so any page can make use
of it without slowly down the page load.

Closes: #1365