jquery-repeat needs to support both active versions of jQuery
Follow up to a199194 - the jquery.repeat plugin breaks with old
versions of jQuery
e.g. author edit page

This is because I didn't understand when writing that patch that
elements will be added to the interface after the code has run and
the document is ready.

Restore live, and its more modern counterpart in jQuery
1.11.0
Fixes: #1396