Goodbye read-btn (conflicts with btn--large)!!
Read eBook button no longer needs to be a "read-btn"
This class is making the button appear smaller than it needs to be now
that it's using btn--large now (the standardised button style going forward)

The works page is the only other consumer of the read-btn and that looks
like a btn primary.

There are no references to read-btn in JS
so goodbye read-btn!

Fixes: #1499