Descope the subjects page
The JavaScript on openlibrary has fallen into disrepair and
the subject page has various functionality that is non-trivial
(in need of a complete rewrite) to work going forward on our "v2"
mode - which modernises jQuery, replaces jCarousels with slick
carousels and mobile friendliness.

This patch descopes a lot of the functionality of the subject pages
specifically it removes the ability to load different carousels by clicking on
sort facets in the graph and in links in the header.

The descoped subject page is trivial to migrate to v2, and when in the future
when our JavaScript is in a better state, it can be restored.

Fixes: #1319