Components: search
A few changes have been made to the CSS:
* .resultTitle h3.booktitle only matches on the search
pages. Other pages which use booktitle do not match resultTitle, so
css rule can be expressed in terms of the new search-result-item component
* searchResultItemCTA is not going to be needed going forward, so although
right now this is used to style this component it will not do so in future.
A general padding/margin rule for direct descendants of search-result-item
will be used in its place.
* A new more descriptive class matching the component name will be used
going forward (search-results-container)

See: #1092