Descope language page and make it v2
it seems that in production we don't have
`languages` in the list of enabled `features` in
the openlibrary conf file
This means master does not reflect production.

When we do this, the page becomes much more simplified.

The resulting pages are handled in readableurls and are made
v2.

The template languages/view.html uses jCarousel so since it is no
longer being used, we throw it away.

The index page is enabled without the feature flag

Possibly this whole page can be removed, but without knowing much about
the original motivations behind it, I can't make this decision.

This however provides the minimum work needed to migrate the language
page to v2 and unblock removing the jCarousel code.

Fixes: #1576