Remove menu JavaScript
On mobile the only way to reveal a menu is to click it.
These JavaScript event handlers close it. You may have noted
how its really hard to click menus on mobile (and sometimes desktop)
and this code appears to be the reason why.

The code doesn't seem to do anything useful. It looks to me like it
was written back in a time when these menus were not implemented via
CSS.

Fixes: #1575