Authors page v2 (#1609)
* Limit some layout styles to desktop breakpoint

* Drop inline styles on author edit form

* Introduce olform__input--large for large textareas such as
the bio field

* Remove some non-standard styles for the author form

Resulting form looks pretty much the same

* Content-quarter moved to legacy and make mobile friendly

This rule has nothing to do with the link-box
component, so should be in legacy.less until we've worked out
what to do with it.
It should be scoped to desktop like the others.

* Rewrite a table based layout on the author edit form

The section
`Please connect Open Library records to good* online resources.`
on /authors/OL27173A/Avi/edit is now consistent with
the section for date of birth/death.

Instead of using tables which makes it difficult to optimise for
mobile it uses a div based layout.

* Textareas should not specify "cols" attribute

For our textareas to be responsive their widths should be defined
via CSS and use media queries. There is no place for cols going
forward!

* Remove 2nd font-size declaration for desktop

We already specified earlier in the stylesheet that the font-size of
a textarea should be 1em.

* Prepare author editor form for async JavaScript

* Use responsive width for certain elements on desktop

This width is too large for various elements on the author edit form.
Instead of using widths defined in pixels make them percentage based
on the containing element

We update the widths for inputs inside olform and the wmd-button-bar

* Prepare ui-tabs for v2

* In updated version of jQuery UI the class changes for active state.
Prepare for this by supporting both classes for active state.
* Make the padding responsive - the large amount of padding is unsuitable
for mobile screens

* Box-sizing border box everywhere!

The relationship between padding and width is painful. In various
places in the UI on mobile, content spills out of its container
as it uses width: 100% and padding. This rectifies this by making
this the default for all v2 pages

* Author page is v2

Fixes: 1578