Component: Lists Page Cta and ListEntry (+ new lists entry point)
Identify two components used on the Lists page.

Since these component are only used on the lists page, create a new
CSS entry point for this page (page-lists.less) that extends the existing
page-user.

This minimises the CSS we ship to other "user" pages.

A #lists h2 selector matches nothing, so remove it.