Fix link adding on author and book edit pages
In a1991944062a7f6daf4b4703476038e0e0eecda0 I fixed a broken selector
which surfaced a new bug. Looking at this some more it is likely
trying to clear the input fields for the link so they don't get
added multiple times. Fix how this works by making it a side effect
of formdata in lieu of a big rewrite.

Add some documentation so the next person understands a bit better
what this is trying to do.

Fixes: #1663