All pages are v2
Maintaining both v2 and v1 is becoming a bit of a burden
and resulting in various issues (such as analytics code which
expects a more modern version of jQuery).

I'm pretty confident at this point that we can deploy the
"nuclear" option and make v2 the default.

v2 boolean flags are left in tact as I'm not sure if they serve
any other purposes (for example in API requests)

"Mr President, it is not only possible, it is essential."
Dr Strangelove or: How I learned to stop worrying about making
Open Library mobile friendly

While this makes Open Library fully responsive, there are lots of
opportunities to improve the UI further in mobile.. there's a long
road ahead!

Untested but should also fix: #1728 #1727

Fixes: #679