Limit work editionCover styling to desktop with margin
In c8595524 I removed the margin right of the editionCover.
Turns out this was needed.

The rule is restored, but only for desktop resolution, as on tablet's
there is not enough space for a 3 column layout - only two.

In process work--tablet is now loaded in work.less rather than
directly from page-book to help discoverability.