Legacy stylesheets are moved out of components
There are 3 files in the components directory which are not yet
components. These are pulled into the main folder and prefixed
legacy. A docblock is added at the top of the files explaining
their status.

The dataTables style which is only needed for JS, is moved into
the js-all entrypoint.