Make the print-disabled-download button a component of read-panel
Introduce a new BEM block class parameter for macros generating
the print disabled button.
The block can be passed as a parameter in case we need to change
it at a later date - either making it a standalone component or
pulling out a new component

For now, it is part of the read-panel component and nothing else.

This means we can remove the helper from misc (and the import statement
inside work). This class cannot really be considered a helper
as it is only used on a handful of pages. Helpers are meant for
general component modifiers that perform single functions e.g.
changing text size or color.