Components must use simpler CSS selectors
To help with our adoption of BEM and to force us to write
more maintable CSS going forward styles in the components
directory will be subject to tighter rules around CSS
specificity and nesting. The .stylelintrc config in this folder will extend
the one in the root folder but will restrict things further

From now on:
* Ids are banned - header-bar styles now make use of `header` selector
since header tag is only ever used in one place in openlibrary
* olBlog component can be removed in favor of some general rules
in home.less that apply to home-resource-section
* Drop various tag name qualifiers on rules
** form.olform can be simplified to .olform
** fieldset.major -> .major
* Disable this rule on certain selectors - we will clean those up
later.

Additional changeS:
* The editions-desktop file is removed  and its contents added
directly to editions.less