Remove unused JS in lazy.js
All the external libraries inside lazy.js are unused in both
templates and JS or duplicated in vendor. Let's remove them or use
the vendor version

Changes:
* .hoverIntent not invoked anwywhere inside openlibrary directory
* No usages of .scrollTo, .highlight, .columnize, .jTruncate
* DataTable plugin (https://www.datatables.net/) is now loaded
from the identical file inside the vendor folder instead of lazy.js
* The tweaks for the dataTable plugin are poorly documented and don't appear
to add any value to the pagination or sorting experience. I have removed it. The
behaviour before and after are equally jarring. Once JS is more under control
I suspect we can do a much better job with this.

Fixes: #1742