Jonrefactors (#1784)
* Legacy stylesheets are moved out of components

There are 3 files in the components directory which are not yet
components. These are pulled into the main folder and prefixed
legacy. A docblock is added at the top of the files explaining
their status.

The dataTables style which is only needed for JS, is moved into
the js-all entrypoint.

* Make the print-disabled-download button a component of read-panel

Introduce a new BEM block class parameter for macros generating
the print disabled button.
The block can be passed as a parameter in case we need to change
it at a later date - either making it a standalone component or
pulling out a new component

For now, it is part of the read-panel component and nothing else.

This means we can remove the helper from misc (and the import statement
inside work). This class cannot really be considered a helper
as it is only used on a handful of pages. Helpers are meant for
general component modifiers that perform single functions e.g.
changing text size or color.

* Components must use simpler CSS selectors

To help with our adoption of BEM and to force us to write
more maintable CSS going forward styles in the components
directory will be subject to tighter rules around CSS
specificity and nesting. The .stylelintrc config in this folder will extend
the one in the root folder but will restrict things further

From now on:
* Ids are banned - header-bar styles now make use of `header` selector
since header tag is only ever used in one place in openlibrary
* olBlog component can be removed in favor of some general rules
in home.less that apply to home-resource-section
* Drop various tag name qualifiers on rules
** form.olform can be simplified to .olform
** fieldset.major -> .major
* Disable this rule on certain selectors - we will clean those up
later.

Additional changeS:
* The editions-desktop file is removed  and its contents added
directly to editions.less