Component CtaBtn: Consolidate cta-btn rules (#1823)
Styling a cta-btn is complicated. This minimises the number of
classes needed and changes the classes to use BEM notation

It should be easier going forward. While it was tempting to create
a macro, it didn't seem to add much given the fact that cta-btn's can
be inputs or anchor tags and thus have various other attributes.