Improve search user experience - Searching should be less jarring
Instead of emptying the results before doing a search, apply an opacity
and empty them just before rendering the results.

Since search is asynchronous this is much less jarring to the user.
Also, clicking anywhere outside the search should close it.

Additional changes:
* Replace deprecated $.ready with $(fn)
* DRY up some CSS selectors so they are only located once
* Update the selector for the search input so it does not apply
to other inputs such as the submit button.