Define a browser support matrix and allow transpiling
We will use Wikipedia's browser support matrix as it is based
on traffic for Wikipedia. Given Wikipedia's traffic, I think we
can be confident this will also be suitable for OpenLibrary.

Add a transpiler to allow us to adopt ES6 syntax in the codebase.
The addition of the babel loader and browserslistrc ensures that the
code we ship to end users will always be suitable syntax for those
users. Note we must still be careful when adopting ES6/7/8 functions
as no polyfills are provided, so care must still be taken by frontend
developers working with this codebase. We can add protection via
eslint later if that becomes a problem.

A single variable is changed to a `const` which does not appear
in the built assets in static/build to demonstrate this is working.

note the minor bundlesize update (but remember we just shaved 5kb off!)

Fixes: #1180