Colorboxlists (#1997)
* Show create list box explicitly

Running $.colorbox will display a colorbox. There is no reason
to bind it to a jquery object unless you are using it to populate
the overlay which we are not doing

See http://www.jacklmoore.com/colorbox/ for description of this
library

Fixes: #1970

* Revert "Revert "Cleanup/document dropclicker behaviour" (#1996)"

This was a red herring and those changes make the code more
readable. The more readable our code is the easier it will
be to maintain

This reverts commit 4e24f1555b8694a1efce4302854c0133ba657b45.

* Functions are exposed with the wrong names

There are references to ungettext and uggettext across our
code without the underscore. This is preventing lists from being
created.

* Render colorbox for list creation so close icon doesn't spill out

This is a follow up to 5e0259071f but generalises that approach for
all elements inside the colorbox component