Jest (#2002)
* Eslint all JS files in the repo

There are other JS files around the repo so let's lint those
as well.

* Add unit testing via the Jest framework

From now on we'll write unit tests to strengthen our code.
I begin by writing tests for two simple files.
In the process, I discover a bug in htmlquote (which I fix!)
Code coverage is reported and lowest coverage possible is defined
in package.json. We will work towards increasing this with time
as we refactor. For now code coverage is 1%.. you have to start
somewhere!!

Fixes: #1999