merge_auuthors correct status and detailed response to help troubleshooting author merge problems relates to #2248 The UI does not test for the success of this request, because the js is now v2 and async with #1609.
I'll need to join up the response, but at least the js has something to sensible to look for. Previously this error was not properly handled, causing a 500 HTTP reponse with no details, so all info was lost, and the 500 error was swallowed by the view page.

Now the JSON error details are at least visible in browser network tab on the
POST to merge.json

This relates to the author's view page, with parameters
/authors/OL..A?merge=true&duplicates=OL..A

example response on error (for the example in #2248):

```
{'message': 'expected /type/author, found /type/redirect', 'at': {'property': 'authors', 'key': '/books/OL11122403M'}, 'value': '/authors/OL47923A', 'error': 'bad_data'}
```