Separate admin.js from standard user JS
Various code is restricted to admins and hidden inside inline templates.
I'd like to make these more visible, testable and reusable by moving them
into an admin only JavaScript entry point.

Focusing on the inline scripts on /admin and /admin/imports seems like a good
start.

Longer term, we may want to move $.plot library into this entry point but
this will require moving graphs on the home page, publishers and
subject page to an admin only view and is out of the scope of this patch.

A modest saving is made on the all.js which no longer needs the
plot_tooltip_graph which is exclusively used on admin pages.