Only load data tables when necessary
The inline script from openlibrary/templates/type/work/editions_datatable.html
is moved into a new entry point openlibrary/plugins/openlibrary/js/edition-tables.js
the code is exactly the same, with the only changes being styling so it confirms to
our eslint standard.

The index.js entry point will now load this file CONDITIONALLY if an
loaded where it is not needed.

To support dynamic importing a babel plugin is added.
"@babel/plugin-syntax-dynamic-import adds support for the import
function

The babel-eslint parser is needed to avoid eslint being
unable to parse this page due to the unexpected import at the bottom
of the file.

Fixes: #2315