Dead code elimination: The dashboard is dead code
When admins view the admin page, a request is made to
https://openlibrary.org/admin/ipstats but nothing is done with the
result... It's supposed to render graphs in element placeholder and overview
but after checking with Mek this never happens.

```
$('#placeholder').length === 0
$('#overview').length === 0
```

Grepping html and JS code there is never an element with ID
tooltip created so associated CSS is also dead.