jQuery UI library is now conditional
We use jQuery UI for a handful of things - changing covers, opening
dialogs and tabs. We don't have these on all pages, so this change
sets up temporary placeholders that load the additional code
only when needed.

A future patch will limit the loading of this library to user actions.
For instance dialogs, tabs are invoked on works page for cover management
when they only need to be loaded when the button is clicked.

Fixes: #2340