Dependabot: Automated Python dependency updates
A great many of our Python dependencies are out-of-data and/or insecure.  Manually updating them is a pain.

Now that the free service [Dependabot](https://dependabot.com) has been acquired by GitHub, should we use it to automate the process open individual pull requests to update each out-of-date dependency so that we can check which ones pass our tests?