Drop unused entry point and templates for login and password
Openlibrary uses Archive.org for account management.
Defer to infogami for vanilla installs (if those are even supported)

The file account.js can be removed as there are no form elements with
the classes email or password and since these functions return early when
that's the case this code is serving no purpose.

Fixes: #2502