Selected shelf on "want to read" is deterministic
Removing a book from a shelf should use a different option
value from the other shelves - otherwise it's not clear to the browser
which option value to render as selected.

Right now we're relying on the first result to be the selected one
but this is what the selected property is for!

The template is DRYed up to be more readable and use the selected property
rather than rely on element order.

The API is modified to accept and remove an entry from its current
shelf when the value -1 is passed.

While here I couldn't help but notice that there is an inconsistency in
openlibrary/templates/account/books.html - while we save a reload by the
code in ol.js removing a book from the shelf, the count for shelves is
always wrong after this takes place.

In the interest of simplifying the frontend, the number of books on a given
shelf is no longer shown in the link label so we don't have to worry about
updating the state of that as changes are made. In future we can revist this
when we have a frontend framework that can track state.

Fixes: #829