disable dependabot python checks until we stabilise
closes #2878

We know a large portion of our Python modules are out of date, and that some of them are not even used for current functionality.
Manual review and fixing of the versions that are causing active problems is required at this stage. Currently minor point version updates are more difficult and risky than they should be, if there is not an active reason to update a module, we shouldn't at this stage.

Removing some out of date unused dependencies is preferable to keeping them current without understanding how we use them, e.g.  #2896

When we remove what is not needed and get to a stable point with Python 3, we can re-enable this tool.