Increase solr timeout
The new solr looks like it takes longer for complicated requests; perhaps because of the ~4M new records. Most average queries look to take about the same though. A strict timeout here cause some weird "pipe discontinued" errors on solr, which seem to make it even slower, so increasing the timeout to prevent that as well.