[library-explorer] Disable linting for vue CSS
Right now the stylelint has lots of restrictions about things like "all colors must be variables", but there's currently no way for us to import a color variable from a .less file into a Vue file (I believe). So for now let's just disable it for Vue, and create an issue to fix it. It'll be largely automated, anyways.