fix loans across workers to not cache
In plugins/upstream/borrow.py within borrow endpoint, the last step before we redirect to bookreader auth link (which performs IA login during redirect) is to fetch (what _was_ cached) loan data. If the book didn't exist in the loan data (which is possible, because of caching) then we'd redirect to the book on archive.org w/o performing the bridge authlink login.

NB: Another more performant solution (if this PR blows up production, which it might because book pages currently do a LoanStatus check which includes a hit to user.loans which was previously cached) then we can simply create a cached and uncached version and then have the borrow logic hit the _uncached_ version.