Imports should be at the top of the file (#4534)
page-form and page-edit both provide styles that override
the defaults specific for that page.

These will be cleaned up by the LESS compiler. To ensure that
the styles in this file override any defaults all imports must
be listed first.

This will address the issue with the change cover background not
being white.