feat: Add cache to GitHub actions
We will add cache to our Python tests which will store the entire
virtual environment in which all our dependencies are installed.

The key for the cache will be the OS version and the hash of all the
requirements.txt files (as the glob suggests). This means that if either
the OS or any of the requirements file changes, the cache will be
automatically updated.

Continuing from https://github.com/internetarchive/infogami/pull/161