Create My Books Dropper using generic dropper
This commit add templates and styling for a generic dropper component,
and recreates the My Books dropper using the generic dropper.  The new
My Books dropper is not yet used in the site, as the Javascript has not
yet been updated.

Classes in the generic dropper are temporarily prefixed with `droppe`,
rather than `dropper`, in order to avoid namespace collisions with the
legacy dropper, and will be updated before release.