Remove active lists showcase
The active lists showcase, which appeared below the dropper on book
pages, has been removed.  Items can now be added and removed from the
dropper itself.  If an item is on a list, a checkmark will be present
next to the dropper list affordance.

Removing a book from the book page "Lists" showcases continues to work,
and correctly updates the dropper list affordances.