Beginning filter edits on team page
Removed bios from the html and changed the sizes of the containers to match.
Added in length check to name -- if someone's name is over a certain character limit, the font size will shrink.
Changed some of the keys in the team.json file.
Began editing the route in code.py to organize team members based on role (staff/fellow/volunteer).