Filtering is working and showing changes on the UI
Filtering works! If the filtering doesn't return any results, a message is appended to the container reflecting that.

On team.json, changed 'role' to 'roles' and 'department' to 'departments', since people can have more than one of each.

On team.less, added the css for the separator between sections -- need to get feedback on what it should look like / if we want to do something else instead.

On code.py and index.html, simplified the python code and the imports