Re-formatting filtering
Changed json roles and departments to be arrays, changed filtering on team.js to match a partial string.

Filtering works in all cases except for when user changes both role and department -- 'fellows' are still showing up in the 'volunteer' section. Should be a quick fix.

After that, still need to put in 'current/past fellows' section and same for volunteers.