feature add ids to works
fixes #3430 and #1797

Adds the option to add external ids (wikidata, VIAF) to works, based
on the code for editions, with minor tweaks.

Will autodetect the id type when possible, similar to functionality
in the author id entry.

Will accept urls for ids and automatically extract the id from them
based on regex sourced from wikidata.