Added filtering to team page (#8437)
* Beginning filter edits on team page

Removed bios from the html and changed the sizes of the containers to match.
Added in length check to name -- if someone's name is over a certain character limit, the font size will shrink.
Changed some of the keys in the team.json file.
Began editing the route in code.py to organize team members based on role (staff/fellow/volunteer).

* Got the filter working in the console

Have it working using a script tag, and only in the console -- nothing is changing on the screen itself yet.

* working on connecting js

Got the watch:js to work by editing the package.json. Still have a couple of errors in the console, one is in the footer component which I haven't touched and the other is a missing jquery module. Hopefully once those get straightened out the js will work!

* Filtering is working and showing changes on the UI

Filtering works! If the filtering doesn't return any results, a message is appended to the container reflecting that.

On team.json, changed 'role' to 'roles' and 'department' to 'departments', since people can have more than one of each.

On team.less, added the css for the separator between sections -- need to get feedback on what it should look like / if we want to do something else instead.

On code.py and index.html, simplified the python code and the imports

* Edited CSS, minor edits to JS

Changed the layout to be more mobile-friendly, made the icons a bit smaller

* Re-formatting filtering

Changed json roles and departments to be arrays, changed filtering on team.js to match a partial string.

Filtering works in all cases except for when user changes both role and department -- 'fellows' are still showing up in the 'volunteer' section. Should be a quick fix.

After that, still need to put in 'current/past fellows' section and same for volunteers.

* Updated team.json, fixed linting errors and filtering

Filtering works -- just need to add in maybe a subtitle and sort for staff current/emeritus and fellows current/past. Also want to keep working on the sorting logic -- it seems inefficient the way I'm doing it.

* Added sub-sections for staff/fellow, fixed team.json

Staff is now separated into "Current" and "Emeritus", Fellows is separated into "Current" and "Past".

Fixed a bug where the "favorite book" icon was showing up on people who didn't have a link to their book in the JSON. Added new Fellows to the JSON, removed bios, and removed some placeholder text.

Filtering all works, but the code seems inefficient. Also need to look at CSS for select tags

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Mek <michael.karpeles@gmail.com>