Use media queries to set `.mybooks-menu` width
This commit uses media queries to set the `.mybooks-menu` width so that
screet widths > @width-breakpoint-tablet and <=
@width-breakpoint-desktop will be set to 200px rather than 250px.
Desktops will be 250px, and phones remain unchanged.

It turned out the overflow issues on tablets went beyond the reading log
carousels; the entire right side of the reading log was overflowing
because the 250px width of `.mybooks-menu` and the 75% width of
`.mybooks-carousel` was just too wide.