Render table asynchronously
Updates loading indicator template to take an optional `data_attrs`
`dict`, which, if present, will be added to the root element as
`data-additional-data`.  This is used to pass the `edition_key` and
`work_key` to the `/partials` handler, which in turn fetches the models
and returns the rendered editions table.