Fix: ensure /api/books defaults to high_priority=False
This commit ensures `/api/books/` passes `high_priority=False`, even if
the endpoint called is _not_ JSON. It appears that this API returns JSON
even if one requests `/api/books` and not `/api/books.json`, and that
conumers do so call the endpoint this way.

Previously, the endpoint only set `high_priority=False` as a default
when accessing the JSON endpoint specifically.