Feature: supplement B* ASIN promise items with BookWorm metadata
This takes advantage of
`/isbn/<asin>?high_priority=true&stage_import=false` on BookWorm to
return metadata without staging anything for import so that the new
metadata can supplement the import record associated with the `B*` ASIN
(e.g. BWB promise item).

The strategy this uses is quite slow, as it goes through the existing
`_get_amazon_metadata`, which is not `async`, and therefore each request
for `B*` metadata takes at least a second. This may need updating in
the future.

Another alternative would be to simply stage the BookWorm records for
import, and update them in `load()`, after first staging them here in
`scripts/promise_batch_imports`