Imports: increase point penalty for non-matching publishers
Edition matching uses a points system found in
`openlibrary/catalog/add_book/match.py`, and a match is made when the
threshold of 875 points is met. This PR increases the penalty for a
truthy and yet non-matching `publishers` field so that editions will not
match when they have matching:
1. author;
2. title; and
3. publish_date.

The rationale is that multiple editions can be published by different
publishers in the same year, and all would have the same title, authors,
and publish_date, yet they would be different editions if the truthy
`publishers` field did not match. If the `publishers` field is falsy,
then a match should happen as it would before, though this does create
the risk that the incoming edition could have an ISBN or other
identifier that doesn't match the existing publisher from the original
record; however, this was the case before and hitherto that has not
appeared to be an issue.