Feature: supplement promise items with AMZ+Google Books
This commit causes promise items to stage additional metadata, if found,
from Amazon or Google Books via Book Worm.

It *also* changes the Just In Time mark-staged-a-pending logic, as it
was discovered this can lead to a race condition during import, whereby
both `pending` records, the BWB record and the BookWorm record, are
imported as unique records.

Because of the change in #9440 such that any record that is incomplete
will look for a `staged` record with which it can complete missing
fields, there may not be a need to mark `staged` items as pending at
all.