Read Work identifiers from file
This is taking the same steps applied to editions[1] and authors[2] and
applying it to works. `get_work_config()` is currently not used for
anything, probably because we’re still waiting for an actual interface
to edit the Work identifiers[3], so this is more a preliminary step to
ensure that Works will be handled in the same way that Editions and
Authors are. (And to allow for PRs to add Work identifiers while waiting
for that UI.)

[1] https://github.com/internetarchive/openlibrary/issues/9234
    https://github.com/internetarchive/openlibrary/pull/9483
[2] https://github.com/internetarchive/openlibrary/issues/9666
    https://github.com/internetarchive/openlibrary/pull/9769
[3] https://github.com/internetarchive/openlibrary/issues/3430