Add Library of Congress NAF as Author identifier
Adds Library of Congress Name Authority File (LC NAF) as an Author identifier.

The `label` has been shortened to just “Library of Congress” to
avoid having an obscenely long label in the drop-down lists and
also avoid obscure acronyms (“NAF”) in it that are likely to
cause more confusion than clarity.

The URI identifiers given by LoC themselves start with “http://” but I
have opted to use “https://” for the `url` so patrons etc. won’t need a
redirect or similar to get to the page.

Closes https://github.com/internetarchive/openlibrary/issues/9791