Include OCLC ids in the Add Book page’s dropdown
OCLC ids will usually be available for titles that do not have
an ISBN to identify them and are also used a lot in Wikipedia
to reference publications.

This uses the same ordering used in the `$popular` array for
adding identifiers on the book editing page.

Fixes https://github.com/internetarchive/openlibrary/issues/8719