Fixed the issue#9946 (#10017)
* Fixed the issue#9946

The issue#9946 was occurring due to "?mode=ebooks" being default at some point in time earlier. Since that is not the case anymore, according to the openlibrary/openlibrary/templates/type/edition/view.html, the editions_limit was being assigned the value 10 always. This prevented the default value of editions_limit to be None (which is then used in  openlibrary/openlibrary/templates/type/work/editions_database.html).

The change includes the 'None' to be one of the acceptable states, restoring the originally intended logic.

* Revert "Fixed the issue#9946"

This reverts commit 5f45188c4b64e6255b40b7502f4e33128c9ac923.

* Fixing editions_datatable.html conditional statement