Support HTTP Proxy on BookWorm
Closes #10230

This commit allows BookWorm to use the HTTP proxy when fetching metadata
from Amazon.

BookWorm relies on an Amazon API client that itself relies on `urllib3`
which explictly does not and will not support the HTTP_PROXY environment
variable.

`urllib3` does support HTTP Proxies, but it needs a specific object, and
the underlying code for the API does support all of this, but the
library provides no way to pass the proxy value through to `urllib3`.

The workaround here is to create the `Configuration` object in the
Amazon API client, add a proxy object after init, then use that to
create a REST client, and then replace the original `api_client` object
with this one, so that there is proxy support.

Note: if trying to create an `vendors.AmazonAPI` object in a REPL, one
will need the 'magic incantation' with `infogami._setup()` after the
config has been loaded, or the `http_proxy_url` value will be `None`.